Move e2e tests into test/e2e package
Restructure for many tests: shared TestMain + runIOS helper in
main_test.go, one test file per command (list_test.go to start).
