+++ .github/workflows/real-device.yml
@@ -15,4 +15,4 @@ jobs:
           cache: false
 
       - name: Run e2e tests
-        run: go test -tags=e2e -count=1 -v .
+        run: go test -tags=e2e -count=1 -v ./test/e2e/...
